docker/factorio-exporter/requirements.txt
factorio-rcon-py>=0.1.4
prometheus-client>=0.20.0
